Refactor AI trip flow into clean controllers, requests, and service
