+++ app/Http/Controllers/AiTripController.php
@@ -4,118 +4,18 @@ namespace App\Http\Controllers;
 
 use App\Enums\Category;
 use App\Models\Activity;
-use App\Models\DayActivity;
 use App\Models\Destination;
 use App\Models\Hotel;
 use App\Models\Trip;
-use App\Models\TripDay;
-use App\Models\TripExclude;
-use App\Models\TripHighlight;
-use App\Models\TripImage;
-use App\Models\TripInclude;
-use App\Models\TripPackage;
-use App\Models\TripPackageHotel;
-use App\Models\TripSchedule;
-use App\Services\GroqTripPlannerService;
-use Illuminate\Http\Request;
-use Illuminate\Support\Facades\DB;
-use Illuminate\Support\Facades\Storage;
-use Illuminate\Support\Str;
 
 class AiTripController extends Controller
 {
-    public function __construct(protected GroqTripPlannerService $groqService)
-    {
-    }
-
     public function create()
     {
         $destinations = Destination::query()->orderBy('name')->get(['id', 'name', 'city', 'country']);
         $categories = Category::cases();
 
         return view('trips.ai.create', compact('destinations', 'categories'));
-    }
-
-    public function generate(Request $request)
-    {
-        $allowedCategories = implode(',', Category::values());
-
-        $validated = $request->validate([
-            'destination_ids' => 'required|array|min:1',
-            'destination_ids.*' => 'required|integer|exists:destinations,id',
-            'description' => 'required|string|max:1000',
-            'categories' => 'required|array|min:1',
-            'categories.*' => 'required|string|in:' . $allowedCategories,
-            'max_participants' => 'required|integer|min:1',
-            'budget' => 'nullable|numeric|min:0',
-            'duration' => 'required|integer|min:1|max:30',
-            'language' => 'nullable|in:en,ar',
-        ]);
-
-        $validated['destination_ids'] = array_values(array_unique($validated['destination_ids']));
-        $validated['categories'] = array_values(array_unique($validated['categories']));
-        $language = $validated['language'] ?? 'en';
-
-        $plan = $this->groqService->generateTripPlan($validated, $language);
-
-        if (! $plan) {
-            return back()->withErrors(['api_error' => 'Failed to generate trip from Groq API.']);
-        }
-
-        $trip = DB::transaction(function () use ($validated, $plan) {
-            $name = Str::limit($plan['trip_name'] ?: $validated['description'], 120, '');
-            $slugBase = Str::slug($name ?: 'ai-trip');
-            $primaryDestinationId = (int) $validated['destination_ids'][0];
-
-            $trip = Trip::create([
-                'destination_id' => $primaryDestinationId,
-                'name' => $name,
-                'slug' => $this->nextUniqueSlug($slugBase),
-                'description' => $plan['trip_description'] ?? null,
-                'duration_days' => (int) $validated['duration'],
-                'category' => implode(',', $validated['categories']),
-                'max_participants' => (int) $validated['max_participants'],
-                'meeting_point_description' => null,
-                'meeting_point_address' => null,
-                'is_ai_generated' => true,
-                'ai_prompt' => $validated['description'],
-                'status' => 'draft',
-            ]);
-
-            $trip->itineraryDestinations()->sync(
-                collect($validated['destination_ids'])->values()->mapWithKeys(fn (int $destinationId, int $index) => [
-                    $destinationId => ['sort_order' => $index + 1],
-                ])->all()
-            );
-
-            foreach ($plan['days'] as $day) {
-                $tripDay = TripDay::create([
-                    'trip_id' => $trip->id,
-                    'day_number' => (int) $day['day_number'],
-                    'title' => $day['title'],
-                    'description' => $day['description'],
-                    'highlights' => [],
-                    'hotel_id' => $day['hotel_id'] ?? null,
-                ]);
-
-                foreach ($day['activities'] as $activity) {
-                    DayActivity::query()->updateOrCreate(
-                        [
-                            'trip_day_id' => $tripDay->id,
-                            'activity_id' => (int) ($activity['activity_id'] ?? 0),
-                        ],
-                        $this->dayActivityAttributes($tripDay->id, $activity)
-                    );
-                }
-            }
-
-            $this->bootstrapPackageFromAiPlan($trip);
-
-            return $trip;
-        });
-
-        return redirect()->route('trip.complete.edit', $trip)
-            ->with('success', 'Your AI trip has been generated as draft. Complete the remaining details and save each tab.');
     }
 
     public function show(Trip $trip)
@@ -147,387 +47,5 @@ class AiTripController extends Controller
         $hotels = Hotel::query()->orderBy('name')->get(['id', 'name']);
 
         return view('trips.ai.complete', compact('trip', 'activeTab', 'destinations', 'activities', 'hotels'));
-    }
-
-    public function saveBasics(Request $request, Trip $trip)
-    {
-        $validated = $request->validate([
-            'destination_id' => 'required|exists:destinations,id',
-            'destination_ids' => 'required|array|min:1',
-            'destination_ids.*' => 'required|integer|exists:destinations,id',
-            'name' => 'required|string|max:255',
-            'description' => 'nullable|string',
-            'duration_days' => 'required|integer|min:1|max:30',
-            'category' => 'nullable|string|max:255',
-            'max_participants' => 'nullable|integer|min:1',
-            'meeting_point_description' => 'nullable|string',
-            'meeting_point_address' => 'nullable|string|max:255',
-        ]);
-
-        $destinationIds = array_values(array_unique(array_map('intval', $validated['destination_ids'])));
-        $primaryDestinationId = (int) $validated['destination_id'];
-
-        if (! in_array($primaryDestinationId, $destinationIds, true)) {
-            array_unshift($destinationIds, $primaryDestinationId);
-        }
-
-        $validated['destination_ids'] = array_values(array_unique([$primaryDestinationId, ...$destinationIds]));
-        $validated['destination_id'] = $primaryDestinationId;
-
-        DB::transaction(function () use ($trip, $validated) {
-            $trip->update(
-                collect($validated)
-                    ->except('destination_ids')
-                    ->put('status', 'draft')
-                    ->all()
-            );
-
-            $trip->itineraryDestinations()->sync(
-                collect($validated['destination_ids'])->values()->mapWithKeys(fn (int $destinationId, int $index) => [
-                    $destinationId => ['sort_order' => $index + 1],
-                ])->all()
-            );
-        });
-
-        return redirect()->route('trip.complete.edit', ['trip' => $trip, 'tab' => 'basics'])
-            ->with('success', 'Basics saved successfully.');
-    }
-
-    public function saveDaysActivities(Request $request, Trip $trip)
-    {
-        $validated = $request->validate([
-            'days' => 'required|array|min:1',
-            'days.*.day_number' => 'required|integer|min:1',
-            'days.*.title' => 'nullable|string|max:255',
-            'days.*.description' => 'nullable|string',
-            'days.*.hotel_id' => 'nullable|exists:hotels,id',
-            'days.*.activities' => 'nullable|array',
-            'days.*.activities.*.id' => 'nullable|integer|exists:day_activities,id',
-            'days.*.activities.*.activity_id' => 'nullable|exists:activities,id',
-            'days.*.activities.*.start_time' => 'nullable|date_format:H:i',
-            'days.*.activities.*.end_time' => 'nullable|date_format:H:i',
-            'days.*.activities.*.notes' => 'nullable|string',
-        ]);
-
-        DB::transaction(function () use ($trip, $validated) {
-            foreach ($validated['days'] as $dayPayload) {
-                $tripDay = TripDay::query()->updateOrCreate(
-                    ['trip_id' => $trip->id, 'day_number' => (int) $dayPayload['day_number']],
-                    [
-                        'title' => $dayPayload['title'] ?? null,
-                        'description' => $dayPayload['description'] ?? null,
-                        'hotel_id' => $dayPayload['hotel_id'] ?? null,
-                    ]
-                );
-
-                $sentActivityIds = [];
-
-                foreach (($dayPayload['activities'] ?? []) as $activityPayload) {
-                    if (empty($activityPayload['activity_id'])) {
-                        continue;
-                    }
-
-                    $activity = DayActivity::query()->updateOrCreate(
-                        [
-                            'id' => $activityPayload['id'] ?? null,
-                            'trip_day_id' => $tripDay->id,
-                        ],
-                        $this->dayActivityAttributes($tripDay->id, $activityPayload)
-                    );
-
-                    $sentActivityIds[] = $activity->id;
-                }
-
-                if (! empty($sentActivityIds)) {
-                    $tripDay->activities()->whereNotIn('id', $sentActivityIds)->delete();
-                } else {
-                    $tripDay->activities()->delete();
-                }
-            }
-
-            $canonicalHotelIds = $this->canonicalHotelIdsFromDays($trip->fresh('days'));
-            $trip->packages()->with('packageHotels')->get()->each(function (TripPackage $package) use ($canonicalHotelIds) {
-                $payload = $package->packageHotels
-                    ->map(fn (TripPackageHotel $packageHotel) => [
-                        'hotel_id' => $packageHotel->hotel_id,
-                        'room_type' => $packageHotel->room_type,
-                        'meal_plan' => $packageHotel->meal_plan,
-                        'amenities' => is_array($packageHotel->amenities) ? implode(', ', $packageHotel->amenities) : '',
-                        'notes' => $packageHotel->notes,
-                    ])
-                    ->values()
-                    ->all();
-
-                $this->syncPackageHotelsFromDays($package, $payload, $canonicalHotelIds);
-            });
-        });
-
-        return redirect()->route('trip.complete.edit', ['trip' => $trip, 'tab' => 'days'])
-            ->with('success', 'Days & activities saved successfully.');
-    }
-
-    public function savePackages(Request $request, Trip $trip)
-    {
-        $validated = $request->validate([
-            'packages' => 'nullable|array',
-            'packages.*.id' => 'nullable|exists:trip_packages,id',
-            'packages.*.name' => 'nullable|string|max:255',
-            'packages.*.price' => 'nullable|numeric|min:0',
-            'packages.*.includes' => 'nullable',
-            'packages.*.includes.*' => 'nullable|string|max:255',
-            'packages.*.excludes' => 'nullable',
-            'packages.*.excludes.*' => 'nullable|string|max:255',
-            'packages.*.highlights' => 'nullable',
-            'packages.*.highlights.*' => 'nullable|string|max:255',
-            'packages.*.hotels' => 'nullable|array',
-            'packages.*.hotels.*.hotel_id' => 'nullable|exists:hotels,id',
-            'packages.*.hotels.*.room_type' => 'nullable|string|max:255',
-            'packages.*.hotels.*.meal_plan' => 'nullable|string|max:255',
-            'packages.*.hotels.*.amenities' => 'nullable|string',
-            'packages.*.hotels.*.notes' => 'nullable|string',
-        ]);
-
-        DB::transaction(function () use ($trip, $validated) {
-            $keepPackageIds = [];
-            $canonicalHotelIds = $this->canonicalHotelIdsFromDays($trip);
-
-            foreach (($validated['packages'] ?? []) as $packagePayload) {
-                if (blank($packagePayload['name'] ?? null) && blank($packagePayload['price'] ?? null)) {
-                    continue;
-                }
-
-                $package = TripPackage::query()->updateOrCreate(
-                    [
-                        'id' => $packagePayload['id'] ?? null,
-                        'trip_id' => $trip->id,
-                    ],
-                    [
-                        'name' => $packagePayload['name'] ?? 'Package',
-                        'price' => $packagePayload['price'] ?? 0,
-                    ]
-                );
-
-                $keepPackageIds[] = $package->id;
-
-                $package->includes()->delete();
-                collect($this->normalizeRepeaterInput($packagePayload['includes'] ?? []))
-                    ->map(fn ($line) => trim($line))
-                    ->filter()
-                    ->each(fn ($line) => TripInclude::create(['trip_package_id' => $package->id, 'content' => $line]));
-
-                $package->excludes()->delete();
-                collect($this->normalizeRepeaterInput($packagePayload['excludes'] ?? []))
-                    ->map(fn ($line) => trim($line))
-                    ->filter()
-                    ->each(fn ($line) => TripExclude::create(['trip_package_id' => $package->id, 'content' => $line]));
-
-                $package->highlights()->delete();
-                collect($this->normalizeRepeaterInput($packagePayload['highlights'] ?? []))
-                    ->map(fn ($line) => trim($line))
-                    ->filter()
-                    ->each(fn ($line) => TripHighlight::create([
-                        'trip_package_id' => $package->id,
-                        'title' => Str::limit($line, 255),
-                        'description' => $line,
-                    ]));
-
-                $this->syncPackageHotelsFromDays($package, $packagePayload['hotels'] ?? [], $canonicalHotelIds);
-            }
-
-            $trip->packages()->whereNotIn('id', $keepPackageIds ?: [0])->delete();
-        });
-
-        return redirect()->route('trip.complete.edit', ['trip' => $trip, 'tab' => 'packages'])
-            ->with('success', 'Packages saved successfully.');
-    }
-
-    public function saveSchedules(Request $request, Trip $trip)
-    {
-        $validated = $request->validate([
-            'schedules' => 'nullable|array',
-            'schedules.*.id' => 'nullable|exists:trip_schedules,id',
-            'schedules.*.start_date' => 'required|date',
-            'schedules.*.end_date' => 'required|date',
-            'schedules.*.booking_deadline' => 'nullable|date',
-            'schedules.*.available_seats' => 'nullable|integer|min:0',
-            'schedules.*.price_modifier' => 'nullable|numeric',
-            'schedules.*.status' => 'nullable|string|in:available,full,cancelled',
-        ]);
-
-        DB::transaction(function () use ($trip, $validated) {
-            $keepIds = [];
-
-            foreach (($validated['schedules'] ?? []) as $schedulePayload) {
-                $schedule = TripSchedule::query()->updateOrCreate(
-                    ['id' => $schedulePayload['id'] ?? null, 'trip_id' => $trip->id],
-                    [
-                        'start_date' => $schedulePayload['start_date'],
-                        'end_date' => $schedulePayload['end_date'],
-                        'booking_deadline' => $schedulePayload['booking_deadline'] ?? null,
-                        'available_seats' => $schedulePayload['available_seats'] ?? null,
-                        'price_modifier' => $schedulePayload['price_modifier'] ?? 0,
-                        'status' => $schedulePayload['status'] ?? 'available',
-                    ]
-                );
-
-                $keepIds[] = $schedule->id;
-            }
-
-            $trip->schedules()->whereNotIn('id', $keepIds ?: [0])->delete();
-        });
-
-        return redirect()->route('trip.complete.edit', ['trip' => $trip, 'tab' => 'schedules'])
-            ->with('success', 'Schedules saved successfully.');
-    }
-
-    public function saveImages(Request $request, Trip $trip)
-    {
-        $validated = $request->validate([
-            'cover_existing_path' => 'nullable|string|max:255',
-            'cover_image_file' => 'nullable|image|mimes:jpg,jpeg,png,webp|max:5120',
-            'images' => 'nullable|array',
-            'images.*.id' => 'nullable|exists:trip_images,id',
-            'images.*.existing_path' => 'nullable|string|max:255',
-            'images.*.image_file' => 'nullable|image|mimes:jpg,jpeg,png,webp|max:5120',
-        ]);
-
-        DB::transaction(function () use ($trip, $validated, $request) {
-            $trip->images()->delete();
-
-            $coverImagePath = $validated['cover_existing_path'] ?? null;
-            if ($request->hasFile('cover_image_file')) {
-                $coverImagePath = '/storage/' . Storage::disk('public')->put('trips', $request->file('cover_image_file'));
-            }
-
-            if (! empty($coverImagePath)) {
-                TripImage::create([
-                    'trip_id' => $trip->id,
-                    'image_path' => $coverImagePath,
-                    'is_cover' => true,
-                ]);
-            }
-
-            foreach (($validated['images'] ?? []) as $index => $imagePayload) {
-                $imagePath = $imagePayload['existing_path'] ?? null;
-
-                if ($request->hasFile("images.$index.image_file")) {
-                    $imagePath = '/storage/' . Storage::disk('public')->put('trips', $request->file("images.$index.image_file"));
-                }
-
-                if (blank($imagePath)) {
-                    continue;
-                }
-
-                TripImage::create([
-                    'trip_id' => $trip->id,
-                    'image_path' => $imagePath,
-                    'is_cover' => false,
-                ]);
-            }
-        });
-
-        return redirect()->route('trip.complete.edit', ['trip' => $trip, 'tab' => 'images'])
-            ->with('success', 'Images saved successfully.');
-    }
-
-    protected function nextUniqueSlug(string $slugBase): string
-    {
-        $slug = $slugBase ?: 'ai-trip';
-        $counter = 1;
-
-        while (Trip::query()->where('slug', $slug)->exists()) {
-            $slug = $slugBase . '-' . $counter;
-            $counter++;
-        }
-
-        return $slug;
-    }
-
-    protected function bootstrapPackageFromAiPlan(Trip $trip): void
-    {
-        if ($trip->packages()->exists()) {
-            return;
-        }
-
-        $package = TripPackage::create([
-            'trip_id' => $trip->id,
-            'name' => 'Standard Package',
-            'price' => 0,
-        ]);
-
-        $hotelIds = $this->canonicalHotelIdsFromDays($trip);
-        foreach ($hotelIds as $hotelId) {
-            TripPackageHotel::create([
-                'trip_package_id' => $package->id,
-                'hotel_id' => $hotelId,
-            ]);
-        }
-    }
-
-    protected function canonicalHotelIdsFromDays(Trip $trip): array
-    {
-        return $trip->days()
-            ->orderBy('day_number')
-            ->whereNotNull('hotel_id')
-            ->pluck('hotel_id')
-            ->map(fn ($hotelId) => (int) $hotelId)
-            ->unique()
-            ->values()
-            ->all();
-    }
-
-    protected function syncPackageHotelsFromDays(TripPackage $package, array $hotelPayloads, array $canonicalHotelIds): void
-    {
-        $payloadByHotelId = collect($hotelPayloads)
-            ->filter(fn ($payload) => ! empty($payload['hotel_id']))
-            ->keyBy(fn ($payload) => (int) $payload['hotel_id']);
-
-        $payloadByIndex = collect($hotelPayloads)->values();
-
-        $package->packageHotels()->delete();
-
-        foreach ($canonicalHotelIds as $index => $hotelId) {
-            $matchedPayload = $payloadByHotelId->get($hotelId) ?? $payloadByIndex->get($index, []);
-
-            TripPackageHotel::create([
-                'trip_package_id' => $package->id,
-                'hotel_id' => $hotelId,
-                'room_type' => $matchedPayload['room_type'] ?? null,
-                'meal_plan' => $matchedPayload['meal_plan'] ?? null,
-                'amenities' => collect(explode(',', (string) ($matchedPayload['amenities'] ?? '')))
-                    ->map(fn ($item) => trim($item))
-                    ->filter()
-                    ->values()
-                    ->all(),
-                'notes' => $matchedPayload['notes'] ?? null,
-            ]);
-        }
-    }
-
-    /**
-     * Keep trip activity timing/notes scoped to day_activities only (independent from master activities table).
-     */
-    protected function dayActivityAttributes(int $tripDayId, array $activityPayload): array
-    {
-        return [
-            'trip_day_id' => $tripDayId,
-            'activity_id' => (int) ($activityPayload['activity_id'] ?? 0),
-            'start_time' => $activityPayload['start_time'] ?? null,
-            'end_time' => $activityPayload['end_time'] ?? null,
-            'notes' => $activityPayload['notes'] ?? null,
-        ];
-    }
-
-    protected function normalizeRepeaterInput(mixed $value): array
-    {
-        if (is_array($value)) {
-            return array_values($value);
-        }
-
-        if (is_string($value)) {
-            return explode(PHP_EOL, $value);
-        }
-
-        return [];
     }
 }
